moved hydrax into separate directory
